move discover screen into seperate pure component.
